// src/Designer/frontend/packages/schema-model/src/types/ReferenceNode.ts
import type { ObjectKind } from './ObjectKind';
import type { KeyValuePairs } from 'app-shared/types/KeyValuePairs';

export type ReferenceNode = {
  objectKind: ObjectKind.Reference;
  schemaPointer: string;
  reference?: string;
  custom: KeyValuePairs;
  description?: string;
  isArray: boolean;
  isNillable: boolean;
  isRequired: boolean;
  title?: string;
  restrictions: KeyValuePairs;
  implicitType: boolean;
};